minor improvements

+++ pkg/shared/utils/xp.go
@@ -11,12 +11,12 @@ func CalculateExperiencePointsFromStats(
 	reactionsReceivedWeight float64,
 	timeSpentVCWeight float64,
 	timeSpentMusicWeight float64,
-) int {
-	var totalExperience int = int(float64(messagesSent)*messageWeight +
+) float64 {
+	var totalExperience float64 = float64(messagesSent)*messageWeight +
 		float64(slashCommandsUsed)*slashCommandWeight +
 		float64(reactionsReceived)*reactionsReceivedWeight +
 		float64(tsVc)*timeSpentVCWeight +
-		float64(tsMusic)*timeSpentMusicWeight)
+		float64(tsMusic)*timeSpentMusicWeight
 
 	return totalExperience
 }
